chore: Extract VssNodesView + VssPathsView from DataBrokerView
